Abstract out `DomainBlockProcessor`
